Status:
1. Finished up a large part of AddressField.java while making field.address
   (App::Dialog::Field::Address).  Having problems and want to talk to you
   to find out what's going on.

2. Added sql and ui directories to CVS repository
3. Converted field.select with "Yes;No" options to a field.boolean
4. Looked at the difference between hint and hints and it seems that hint is
   a. used in one or two places at most, and
   b. hints is most often used for text that provides clues to the user about
      the field while hint seems to be used to provide default values or to
      check against default values in a field.
5. Updated WIP with the contents we discussed yesterday.

FYI, most of my time was spent getting 1 up and running.  I've gotten complete
validation for state and zip codes which are translations of the validation
routines from ASPire.  The field.address gets registered in the dialog and
shows up in the FieldFactory view even but is not rendered properly.

Perhaps I'm doing something wrong in the way I'm creating the AddressField
object?